Clearing post-user relations

+++ controllers/feed.js
@@ -116,12 +116,17 @@ exports.updatePost = (req, res, next) => {    const errors = validationResult(re
     }
     Post.findById(postId)
         .then(post => {
-            if(!post) {
+            if (!post) {
                 const error = new Error('Could not find post.');
                 error.statusCode = 404;
                 throw error
             }
-            if(imageUrl !== post.imageUrl) {
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not autorized!');
+                error.statusCode = 403;
+                throw error;
+            }
+            if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
             post.title = title;
@@ -149,13 +154,24 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not autorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             // check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Deleted post.' })
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
+            res.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
             if (!err.statusCode) {
